:construction_worker: Use `conan build` for package testing
Conan folks have stated that testing should be apart of building and
that using conan create and directing the `test_package` to the `test/`
directory is bad form. This added additional complexity and confusing
arguments to command in that it disabled tests and yet it was the
tests.yml.

Instead this change uses `conan build` to build the package and store
the results in a `build/` directory at the package's root. This makes
the process much more clear, conforming to the conan best practices and
simpler to understand for readers.

- Format a few of the .yml files
- Remove "test disable" from build