:sparkles: Add platform_deploy.yml
The previous CI used the host settings for the builds. This works, but
this would mean that package consumers for bare metal packages have to
build them from source on their host machines. The goal is to ensure
that users never have to build from source on their host machines and
can simply use the prebuilt binaries made available in the artifactory
repository. This workflow works for both platforms and processor
libraries.

This change:

- Removes build.yml from library.yml ci
- Adds a platform_deploy.yml which install the bare metal settings,
  installs the profiles in the conan/profiles/ directory, builds the
  package based on the profile set in the input, and tests that the
  demos work with that package. The approach allows repo managers to
  create multiple parallel jobs for each profile that a platform or
  processor uses.
- Adds self_check tests
- Fixes up lint.yml to support checking src/ directories